Fixed Single Policy Trainer
