+++ abmarl/examples/__init__.py
@@ -1,4 +1,5 @@
 from .sim import MultiCorridor
 from .sim import BattleAgent, TeamBattleSim
 from .sim import MazeNavigationAgent, MazeNaviationSim
-from .sim import EmptyABS, MultiAgentSim, MultiAgentGymSpacesSim, MultiAgentContinuousGymSpaceSim
+from .sim import EmptyABS, MultiAgentSim, MultiAgentGymSpacesSim, \
+    MultiAgentContinuousGymSpaceSim, MultiAgentSameSpacesSim
