+++ abmarl/examples/sim/__init__.py
@@ -2,4 +2,4 @@ from .multi_corridor import MultiCorridor
 from .team_battle_example import BattleAgent, TeamBattleSim
 from .maze_navigation import MazeNavigationAgent, MazeNaviationSim
 from .multi_agent_sim import EmptyABS, MultiAgentSim, MultiAgentGymSpacesSim, \
-    MultiAgentContinuousGymSpaceSim
+    MultiAgentContinuousGymSpaceSim, MultiAgentSameSpacesSim
