+++ tests/trainers/test_single_policy_trainer.py
@@ -2,92 +2,12 @@
 from gym.spaces import Discrete, MultiBinary, Box, Dict, Tuple
 import pytest
 
-from abmarl.sim.agent_based_simulation import AgentBasedSimulation, Agent, PrincipleAgent
+from abmarl.sim.agent_based_simulation import Agent
 
 from abmarl.trainers import SinglePolicyTrainer
 from abmarl.policies.policy import RandomPolicy
 from abmarl.managers import AllStepManager, TurnBasedManager
-
-
-class SimTest(AgentBasedSimulation):
-    def __init__(self):
-        self.rewards = [0, 1, 2, 3, 4, 5, 6]
-        self.ith_call = -1
-        self.dones = [12, 25]
-        self.step_count = 0
-        self.agents = {
-            'agent0': Agent(
-                id='agent0',
-                observation_space=MultiBinary(4),
-                action_space=Tuple((
-                    Dict({
-                        'first': Discrete(4),
-                        'second': Box(low=-1, high=3, shape=(2,), dtype=int)
-                    }),
-                    MultiBinary(3)
-                ))
-            ),
-            'agent1': PrincipleAgent(
-                id='agent1',
-            ),
-            'agent2': Agent(
-                id='agent2',
-                observation_space=MultiBinary(4),
-                action_space=Tuple((
-                    Dict({
-                        'first': Discrete(4),
-                        'second': Box(low=-1, high=3, shape=(2,), dtype=int)
-                    }),
-                    MultiBinary(3)
-                ))
-            ),
-            'agent3': PrincipleAgent(
-                id='agent3',
-            ),
-            'agent4': PrincipleAgent(
-                id='agent4'
-            )
-        }
-
-    def render(self):
-        pass
-
-    def reset(self):
-        self.action = {
-            'agent0': None,
-            'agent2': None,
-        }
-
-    def step(self, action_dict):
-        self.step_count += 1
-        for agent_id, action in action_dict.items():
-            self.action[agent_id] = action
-
-    def get_obs(self, agent_id, **kwargs):
-        if agent_id == 'agent0':
-            return [0, 0, 0, 1]
-        elif agent_id == 'agent2':
-            return [1, 0, 1, 0]
-
-    def get_reward(self, agent_id, **kwargs):
-        self.ith_call = (self.ith_call + 1) % 7
-        return self.rewards[self.ith_call]
-
-    def get_done(self, agent_id, **kwargs):
-        if agent_id == 'agent0':
-            return self.step_count >= self.dones[0]
-        elif agent_id == "agent2":
-            return self.step_count >= self.dones[1]
-
-    def get_all_done(self, **kwargs):
-        for agent in self.agents.values():
-            if not isinstance(agent, Agent): continue
-            if not self.get_done(agent.id):
-                return False
-        return True
-
-    def get_info(self, agent_id, **kwargs):
-        return self.action[agent_id]
+from abmarl.examples import MultiAgentSameSpacesSim
 
 
 class NoTrainer(SinglePolicyTrainer):
@@ -95,7 +15,7 @@ class NoTrainer(SinglePolicyTrainer):
         return self.generate_episode(horizon=20)
 
 
-sim = AllStepManager(SimTest())
+sim = AllStepManager(MultiAgentSameSpacesSim())
 policy = RandomPolicy(
     action_space=sim.agents['agent0'].action_space,
     observation_space=sim.agents['agent0'].observation_space,
@@ -104,14 +24,14 @@ policy = RandomPolicy(
 
 def test_trainer_sim():
     with pytest.raises(AssertionError):
-        NoTrainer(sim=SimTest(), policy=policy)
+        NoTrainer(sim=MultiAgentSameSpacesSim(), policy=policy)
 
     trainer = NoTrainer(sim=sim, policy=policy)
 
     with pytest.raises(AssertionError):
-        trainer.sim = SimTest()
+        trainer.sim = MultiAgentSameSpacesSim()
 
-    trainer.sim = TurnBasedManager(SimTest())
+    trainer.sim = TurnBasedManager(MultiAgentSameSpacesSim())
 
 
 def test_trainer_policy():
